@@ -8,8 +8,8 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.24.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "18",
+    "react-dom": "18",
     "react-icons": "^4.3.1",
     "react-image-gallery": "^1.2.7",
     "react-query": "^3.34.15",
